+++ src/BlazorCardGame.WebUI/Store/RoundStateCase/RoundState.cs
@@ -3,10 +3,10 @@ public class RoundState
 {
     public int HandsRemaining { get; }
     public int DiscardsRemaining { get; }
-    public int RoundScore { get; }
+    public long RoundScore { get; }
 
     public RoundState() { }
-    public RoundState(int handsRemaining, int discardsRemaining, int roundScore)
+    public RoundState(int handsRemaining, int discardsRemaining, long roundScore)
     {
         HandsRemaining = handsRemaining;
         DiscardsRemaining = discardsRemaining;
